+++ lib/logging/monitor.rb
@@ -1,11 +1,21 @@
 # frozen_string_literal: true
 
-require_relative 'call_location'
-
 module Logging
-  class Monitor
-    include Logging::CallLocation
+  # Proxy class to allow a custom `caller_location` to be used
+  class CallLocation
+    attr_accessor :base_label, :path, :lineno
 
+    # create proxy caller_location
+    # @see Thread::Backtrace::Location
+    # @see ZeroSilentFailures::Monitor#parse_caller
+    def initialize(function = nil, file = nil, line = nil)
+      @base_label = function
+      @path = file
+      @lineno = line
+    end
+  end
+
+  class Monitor
     # create a monitor
     def initialize(service)
       @service = service
